[FIX] separação do res.partner e do res.partner.bank
