+++ l10n_br_base/models/__init__.py
@@ -5,6 +5,7 @@
 
 from . import res_bank
 from . import res_partner
+from . import res_partner_bank
 from . import res_country
 from . import res_company
 from . import l10n_br_base_city
